In this I have fixed an issue regarding the saved tree number.
The problem was that the "Tree Number" we were saving was the task local
tree number; not the original tree number from the input files.

I have fixed this by recording the tree number when setting up the forest information.
This correctly accounts for the fact that we can start reading part-way through a tree file.

Test data will need to be updated as the previous values for "SAGETreeIndex" were incorrect.