Reworked the flow (again, sigh).
We now use the SAGE parameter file directly, reducing the need to specify a whole bunch of stuff.

As a result, the data subclasses should no longer be subclasses. Instead, they're just normal classes who's __init__
method requires a Model class to call them.

Finally, the OutputDir path for the SAGE parameter file is required to be an absolute path now as a result.