Reading the SAGE file is now done at the data class level.
This has the side effect of the user having to specifically specify what
the `OutputFormat` of each model is (e.g., `sage_binary` or
`sage_hdf5`).  Whilst this may appear to be redundant as it could be
read in from the SAGE parameter file, this will allow the integration of
future data formats much easier.

For example, reading in the output of a future SAGE model (e.g., Pipit's
data format) will now be much much easier.